+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to
 
 - Changed incorrectly spelled `github_app` property `respositorySelected` to
   `repositorySelected`
+- Better error handling, including support for when GitHub errors are provided
+  with non-error (200) codes.
 
 ## 1.7.0 - 2021-10-18